Update downloads and errors
