+++ _scripts/previous.js
@@ -8,13 +8,24 @@
 import jQuery from '~/lib/jquery'
 
 import config from '~/config'
+import detectedOS from '~/widgets/detectedos'
 
 Promise.all([config, jQuery]).then(([config, $]) => {
     $(document).ready(() => {
         // ACTION: .download-http.click: Track download over HTTP
         $('.download-link').click(function () {
-            plausible('Download of Previous Version')
-            plausible('Download of ' + config.previous.title + ' ' + config.previous.version)
+            if ( $(this).hasClass('http') ) {
+                var method = 'HTTP'
+            }
+            if ( $(this).hasClass('magnet') ) {
+                var method = 'Magnet'
+            }
+            plausible('Downloads', {meta: {
+                Region: config.user.region,
+                Method: method,
+                OS: detectedOS,
+                Version: config.previous.version
+            }})
         })
         console.log('Loaded previous.js')
     })
